refactor: delete unusable and commented code
