+++ src/app.module.ts
@@ -6,14 +6,12 @@ import { TrackModule } from './track/track.module';
 import { ArtistModule } from './artist/artist.module';
 import { AlbumModule } from './album/album.module';
 import { FavsModule } from './favs/favs.module';
-import { DBModule } from './db/db.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { dbAsyncConfig } from './db/db.config.service';
 import { ConfigModule } from '@nestjs/config';
 
 @Module({
   imports: [
-    DBModule,
     UserModule,
     TrackModule,
     ArtistModule,
