+++ src/artist/entities/artist.entity.ts
@@ -2,8 +2,7 @@ import { HttpStatus } from '@nestjs/common';
 import { IsBoolean, IsString } from 'class-validator';
 import { Album } from 'src/album/entities/album.entity';
 import { appError } from 'src/errors';
-import { Favorites } from 'src/favs/entities/favorites.entity';
-// import { FavArtists } from 'src/favs/entities/fav.artist.entity';
+import { Favorites } from 'src/favs/entities/favs.entity';
 import { Track } from 'src/track/entities/track.entity';
 import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
 import { v4 } from 'uuid';
