--- src/db/db.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { DB } from './db.service';
-
-@Controller()
-export class DBController {
-  constructor(private readonly db: DB) {}
-}
