+++ src/user/user.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { DB } from 'src/db/db.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 
 @Module({
   controllers: [UserController],
   imports: [TypeOrmModule.forFeature([User])],
-  providers: [UserService, DB],
+  providers: [UserService],
 })
 export class UserModule {}
